Moved files to new folder, and changed encoder code to work for PA0 and PA1

+++ ME 507 Clawmachine/src/Michael's edits/encoder_counter.h
@@ -45,7 +45,7 @@ public:
     /** @brief   Return the current position count from the timer.
      *  @returns The value in the timer's count register
      */
-    uint16_t getCount (void)
+    int32_t getCount (void)
     {
         return p_timer->getCount ();
     }
